requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.12.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.4
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.4
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.4.0
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
zope.interface==5.5.2
